refactor(lib): share pygit2 config hardening between bundled scripts
Extract read_global_http_proxy + harden_pygit2_config (the latter now returns
the snapshotted http.proxy) into lib/pygit2_compat.py, imported by both
git_operations.py and update_comfyui.py instead of duplicating them. The whole
lib/ dir ships via extraResources and the scripts run with -s/-u (script dir on
sys.path), so the sibling import resolves in dev and packaged builds; verified
by running both scripts.

Amp-Thread-ID: https://ampcode.com/threads/T-019eafa0-16a1-726e-91a4-dac1a40d4481
